refactor: routes for scriptanswer, scripts and tasks logic now in separate files
