+++ server/src/server_routes/scriptanswer.ts
@@ -1,0 +1,118 @@
+import { Request, Response, Express } from 'express';
+
+export function setupScriptAnswerRoutes(app: Express, scriptAnswerSet: any, studentSet: any) {
+  const scriptAnswerurl = '/api/scriptanswers/';
+
+  // GET /api/scriptanswers → get ALL answers
+  app.get(scriptAnswerurl+'', (req: Request, res: Response) => {
+    try {
+      const all = scriptAnswerSet.getAll();
+      return res.status(200).json(all.map((a: any) => a.toJSON()));
+    } catch (err) {
+      return res.status(500).json({ error: 'Failed to fetch script answers' });
+    }
+  });
+
+  // GET /api/scriptanswers/student/:studentId → Answers for a student
+  app.get(scriptAnswerurl+'student/:studentId', (req: Request, res: Response) => {
+    const { studentId } = req.params;
+
+    const student = studentSet.findStudentByCPF(studentId);
+    if (!student) {
+      return res.status(404).json({ error: 'Student not found' });
+    }
+
+    const answers = scriptAnswerSet.findByStudentId(studentId);
+    return res.status(200).json(answers.map((a: any) => a.toJSON()));
+  });
+
+  // GET /api/scriptanswers/script/:scriptId → Answers for a script
+  app.get(scriptAnswerurl+'script/:scriptId', (req: Request, res: Response) => {
+    const { scriptId } = req.params;
+
+    const answers = scriptAnswerSet.findbyScriptId(scriptId);
+    return res.status(200).json(answers.map((a: any) => a.toJSON()));
+  });
+
+  // GET /api/scriptanswers/:id → get ONE answer by ID
+  app.get(scriptAnswerurl+':id', (req: Request, res: Response) => {
+    const { id } = req.params;
+
+    const found = scriptAnswerSet.findById(id);
+    if (!found) {
+      return res.status(404).json({ error: 'ScriptAnswer not found' });
+    }
+
+    return res.status(200).json(found.toJSON());
+  });
+
+  // GET /api/scriptanswers/:id/tasks/:taskId → Get grade of task
+  app.get(scriptAnswerurl+':id/tasks/:taskId', (req: Request, res: Response) => {
+    const { id, taskId } = req.params;
+
+    const scriptAnswer = scriptAnswerSet.findById(id);
+    if (!scriptAnswer) {
+      return res.status(404).json({ error: 'ScriptAnswer not found' });
+    }
+
+    const task = scriptAnswer.findAnswerByTaskId(taskId);
+    if (!task) {
+      return res.status(404).json({ error: 'Task not found' });
+    }
+
+    return res.status(200).json({ taskId, grade: task.getGrade() });
+  });
+
+  // POST /api/scriptanswers → Create a script answer
+  app.post(scriptAnswerurl+'', (req: Request, res: Response) => {
+    try {
+      const newAnswer = scriptAnswerSet.addScriptAnswer(req.body);
+      res.status(201).json(newAnswer.toJSON());
+    } catch (error) {
+      res.status(400).json({ error: (error as Error).message });
+    }
+  });
+
+  // PUT /api/scriptanswers/:id/tasks/:taskId → Update grade
+  app.put(scriptAnswerurl+':id/tasks/:taskId', (req: Request, res: Response) => {
+    const { id, taskId } = req.params;
+    const { grade } = req.body;
+
+    const scriptAnswer = scriptAnswerSet.findById(id);
+    if (!scriptAnswer) {
+      return res.status(404).json({ error: 'ScriptAnswer not found' });
+    }
+
+    const task = scriptAnswer.findAnswerByTaskId(taskId);
+    if (!task) {
+      return res.status(404).json({ error: 'Task not found' });
+    }
+
+    if (!['MANA', 'MPA', 'MA'].includes(grade)) {
+      return res.status(400).json({ error: 'Invalid grade' });
+    }
+
+    task.updateGrade(grade);
+    return res.status(200).json({ taskId, grade });
+  });
+
+  // PUT /api/scriptanswers/:id/tasks/:taskId/comments → Add comment to task
+  app.put(scriptAnswerurl+':id/tasks/:taskId/comments', (req: Request, res: Response) => {
+    const { id, taskId } = req.params;
+    const { comment } = req.body;
+
+    const scriptAnswer = scriptAnswerSet.findById(id);
+    if (!scriptAnswer) {
+      return res.status(404).json({ error: 'ScriptAnswer not found' });
+    }
+
+    const task = scriptAnswer.findAnswerByTaskId(taskId);
+    if (!task) {
+      return res.status(404).json({ error: 'Task not found' });
+    }
+
+    task.comments = comment;
+
+    return res.status(200).json({ taskId, comment });
+  });
+}
